Sprite reference names
Uploading SVGs with names of existing files makes the `$post->post_name` enumerated, which is probably not the expected result when you want to reference that sprite. For example, you upload `logo.svg`, but there's already a `logo.jpg`, so currently you have to use `logo-2` to reference `logo.svg`'s sprite.

My solution uses a sanitized `$post->post_title`, which defaults to the filename on upload but can be edited by the user if they choose (`logo` to `logo-red` or `logo-horizontal`). Duplicate names within the sprite sheet are enumerated (`logo` and `logo-2`) as necessary.

For existing users, this may cause slight compatibility issues where the `post_title` may be different than the file name, or where they were already referencing the enumerated `post_name`, but this should be a more elegant solution overall.